+++ app/helpers/foreman_dlm/dlmlock_helper.rb
@@ -18,7 +18,7 @@ module ForemanDlm
       if lock.enabled?
         actions << display_link_if_authorized(
           _('Disable'),
-          hash_for_disable_dlmlock_path(:id => lock.to_param).merge(auth_object: lock, authorizer: authorizer),
+          hash_for_disable_foreman_dlm_dlmlock_path(:id => lock.to_param).merge(auth_object: lock, authorizer: authorizer),
           method: :put
         )
       end
@@ -26,7 +26,7 @@ module ForemanDlm
       if lock.disabled?
         actions << display_link_if_authorized(
           _('Enable'),
-          hash_for_enable_dlmlock_path(:id => lock.to_param).merge(auth_object: lock, authorizer: authorizer),
+          hash_for_enable_foreman_dlm_dlmlock_path(:id => lock.to_param).merge(auth_object: lock, authorizer: authorizer),
           method: :put
         )
       end
@@ -34,12 +34,12 @@ module ForemanDlm
       if lock.taken?
         actions << display_link_if_authorized(
           _('Release'),
-          hash_for_release_dlmlock_path(:id => lock.to_param).merge(auth_object: lock, authorizer: authorizer),
+          hash_for_release_foreman_dlm_dlmlock_path(:id => lock.to_param).merge(auth_object: lock, authorizer: authorizer),
           method: :put
         )
       end
 
-      actions << display_delete_if_authorized(hash_for_dlmlock_path(:id => lock.to_param).merge(auth_object: lock, authorizer: authorizer), class: 'delete')
+      actions << display_delete_if_authorized(hash_for_foreman_dlm_dlmlock_path(:id => lock.to_param).merge(auth_object: lock, authorizer: authorizer), class: 'delete')
       actions
     end
   end
