+++ app/views/foreman_dlm/dlmlocks/_details.html.erb
