+++ app/views/foreman_dlm/dlmlocks/show.html.erb
@@ -1,0 +1,10 @@
+<% title @dlmlock.name %>
+<% breadcrumbs(
+   resource_url: api_dlmlocks_path
+) if respond_to?(:breadcrumbs) # Requires Foreman >= 1.18 %>
+
+<%= title_actions link_to(_('Back'), foreman_dlm_dlmlocks_path, :class => 'btn btn-default'),
+  ((link_to(_("Host details"), @dlmlock.host, :class => 'btn btn-default') if @dlmlock.host.present?))
+%>
+
+<%= render 'details' %>
